Upgrade to vue cli 5 and try to add AVIF format
Generating AVIF images technically works, but is a bit pointless unless I can figure out the magic combination of params to make it good.  As of right now it is extremely slow (several minutes per image) and isn't really generating smaller images.  With cqLevel of 10 (out of 63 being crap) it is the same size as a regular PNG with still some minor artifacts.  When we go to lossless compression it is bigger than the PNG unless we use the slowest speed and it takes 15 minutes to run for a tiny bit of saved space.  At this point it is better just to use webp.  I also might need to look into just using sharp instead of trying to rely on browser implementations of libavif since sharp seems to be faster, but is node only.